Address PR review feedback (#15) — small refactors
- ToggleFilter: extract tokenForIndex/indexOfToken/indexOfValue into
  optionTokens.ts for reuse and testability.
- AIDescriptionSection: split Spinner into its own components/ui/Spinner.tsx
  (one component per file).
- MatchGroupTile: extract overlay badge selection into renderOverlayBadge
  helper.
- InstagramTab: extract overlay badge fragment into
  renderInstagramOverlayBadges helper.

Made-with: Cursor
