+++ apps/visualizer/frontend/src/components/DescriptionPanel/AIDescriptionSection.tsx
@@ -6,6 +6,7 @@ import { useJobSocket } from '../../hooks/useJobSocket'
 import { DescriptionPanel } from './DescriptionPanel'
 import { GenerateButton } from '../ui/description-atoms/GenerateButton'
 import { ProviderModelSelect } from '../ui/ProviderModelSelect'
+import { Spinner } from '../ui/Spinner'
 import { IMAGE_DETAILS_AI_DESCRIPTION } from '../../constants/strings'
 
 interface AIDescriptionSectionProps {
@@ -160,25 +161,5 @@ export function AIDescriptionSection({
         />
       ) : null}
     </div>
-  )
-}
-
-function Spinner() {
-  return (
-    <svg className="animate-spin h-4 w-4 text-accent" viewBox="0 0 24 24" fill="none">
-      <circle
-        className="opacity-25"
-        cx="12"
-        cy="12"
-        r="10"
-        stroke="currentColor"
-        strokeWidth="4"
-      />
-      <path
-        className="opacity-75"
-        fill="currentColor"
-        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4z"
-      />
-    </svg>
   )
 }
